+++ src/main/java/com/bon/modules/process/domain/dto/FormExtGroupDTO.java
@@ -8,7 +8,7 @@ import com.bon.common.domain.dto.BaseDTO;
 import com.bon.modules.process.domain.entity.FormExtGroup;
 
 /**
- * @Created：2018-08-27
+ * @Created：2018-08-28
  * @Author Albert
  * @Version: 1.0
  * @Description: FormExtGroup参数类
@@ -28,13 +28,13 @@ public class FormExtGroupDTO extends BaseDTO<FormExtGroup> implements Serializab
     private Date gmtModified;
 
     @ApiModelProperty(value = "表单类型ID")
-    private Integer formTypeId;
+    private Long formTypeId;
 
     @ApiModelProperty(value = "表单TabID")
-    private Integer tabId;
+    private Long tabId;
 
     @ApiModelProperty(value = "表单项组别：;与proc_setting.node_id关联;默认为0;为0时表示流程开始就必须填写;不为0时表示走到相应流程时填写")
-    private Integer nodeId;
+    private Long nodeId;
 
     @ApiModelProperty(value = "唯一标识码")
     private String name;
@@ -69,27 +69,27 @@ public class FormExtGroupDTO extends BaseDTO<FormExtGroup> implements Serializab
         this.gmtModified = gmtModified;
     }
 
-    public Integer getFormTypeId() {
+    public Long getFormTypeId() {
         return formTypeId;
     }
 
-    public void setFormTypeId(Integer formTypeId) {
+    public void setFormTypeId(Long formTypeId) {
         this.formTypeId = formTypeId;
     }
 
-    public Integer getTabId() {
+    public Long getTabId() {
         return tabId;
     }
 
-    public void setTabId(Integer tabId) {
+    public void setTabId(Long tabId) {
         this.tabId = tabId;
     }
 
-    public Integer getNodeId() {
+    public Long getNodeId() {
         return nodeId;
     }
 
-    public void setNodeId(Integer nodeId) {
+    public void setNodeId(Long nodeId) {
         this.nodeId = nodeId;
     }
