+++ src/main/java/com/bon/modules/process/domain/entity/Form.java
@@ -11,7 +11,7 @@ import javax.persistence.GeneratedValue;
 
 
 /**
- * @Created：2018-08-27
+ * @Created：2018-08-28
  * @Author Albert
  * @Version: 1.0
  * @Description: Form参数类
@@ -35,10 +35,10 @@ public class Form implements Serializable{
     private Date gmtModified;
 
     @ApiModelProperty(value = "表单类型ID：与form_type.form_type_id关联")
-    private Integer formTypeId;
+    private Long formTypeId;
 
     @ApiModelProperty(value = "当前所处节点ID")
-    private Integer curNodeId;
+    private Long curNodeId;
 
     @ApiModelProperty(value = "实例名称")
     private String name;
@@ -48,7 +48,7 @@ public class Form implements Serializable{
     private Date createTime;
 
     @ApiModelProperty(value = "建立人")
-    private Integer creatorId;
+    private Long creatorId;
 
     public Long getFormId() {
         return formId;
@@ -74,19 +74,19 @@ public class Form implements Serializable{
         this.gmtModified = gmtModified;
     }
 
-    public Integer getFormTypeId() {
+    public Long getFormTypeId() {
         return formTypeId;
     }
 
-    public void setFormTypeId(Integer formTypeId) {
+    public void setFormTypeId(Long formTypeId) {
         this.formTypeId = formTypeId;
     }
 
-    public Integer getCurNodeId() {
+    public Long getCurNodeId() {
         return curNodeId;
     }
 
-    public void setCurNodeId(Integer curNodeId) {
+    public void setCurNodeId(Long curNodeId) {
         this.curNodeId = curNodeId;
     }
 
@@ -106,11 +106,11 @@ public class Form implements Serializable{
         this.createTime = createTime;
     }
 
-    public Integer getCreatorId() {
+    public Long getCreatorId() {
         return creatorId;
     }
 
-    public void setCreatorId(Integer creatorId) {
+    public void setCreatorId(Long creatorId) {
         this.creatorId = creatorId;
     }
