+++ src/main/java/com/bon/modules/process/domain/entity/FormExt.java
@@ -11,7 +11,7 @@ import javax.persistence.GeneratedValue;
 
 
 /**
- * @Created：2018-08-27
+ * @Created：2018-08-28
  * @Author Albert
  * @Version: 1.0
  * @Description: FormExt参数类
@@ -35,13 +35,13 @@ public class FormExt implements Serializable{
     private Date gmtModified;
 
     @ApiModelProperty(value = "表单实例ID:关联form.form_id")
-    private Integer formId;
+    private Long formId;
 
     @ApiModelProperty(value = "扩展表单项ID：关联form_ext_setting.option_id")
-    private Integer optionId;
+    private Long optionId;
 
     @ApiModelProperty(value = "表单项组别")
-    private Integer optGroup;
+    private Long optGroup;
 
     @ApiModelProperty(value = "表单类型:00：短文本;01：长文本;02：下拉选项;03：多选项;04：单选项")
     private String optType;
@@ -79,27 +79,27 @@ public class FormExt implements Serializable{
         this.gmtModified = gmtModified;
     }
 
-    public Integer getFormId() {
+    public Long getFormId() {
         return formId;
     }
 
-    public void setFormId(Integer formId) {
+    public void setFormId(Long formId) {
         this.formId = formId;
     }
 
-    public Integer getOptionId() {
+    public Long getOptionId() {
         return optionId;
     }
 
-    public void setOptionId(Integer optionId) {
+    public void setOptionId(Long optionId) {
         this.optionId = optionId;
     }
 
-    public Integer getOptGroup() {
+    public Long getOptGroup() {
         return optGroup;
     }
 
-    public void setOptGroup(Integer optGroup) {
+    public void setOptGroup(Long optGroup) {
         this.optGroup = optGroup;
     }
