+++ src/main/java/com/bon/modules/process/domain/entity/Process.java
@@ -11,7 +11,7 @@ import javax.persistence.GeneratedValue;
 
 
 /**
- * @Created：2018-08-27
+ * @Created：2018-08-28
  * @Author Albert
  * @Version: 1.0
  * @Description: Process参数类
@@ -35,10 +35,10 @@ public class Process implements Serializable{
     private Date gmtModified;
 
     @ApiModelProperty(value = "表单实例ID:关联form.form_id")
-    private Integer formId;
+    private Long formId;
 
     @ApiModelProperty(value = "扩展表单项ID：关联form_ext_setting.option_id")
-    private Integer nodeId;
+    private Long nodeId;
 
     @ApiModelProperty(value = "节点名称")
     private String nodeName;
@@ -90,19 +90,19 @@ public class Process implements Serializable{
         this.gmtModified = gmtModified;
     }
 
-    public Integer getFormId() {
+    public Long getFormId() {
         return formId;
     }
 
-    public void setFormId(Integer formId) {
+    public void setFormId(Long formId) {
         this.formId = formId;
     }
 
-    public Integer getNodeId() {
+    public Long getNodeId() {
         return nodeId;
     }
 
-    public void setNodeId(Integer nodeId) {
+    public void setNodeId(Long nodeId) {
         this.nodeId = nodeId;
     }
