+++ src/main/java/com/bon/modules/process/domain/vo/FormExtGroupVO.java
@@ -4,7 +4,7 @@ import java.util.*;
 import java.io.Serializable;
 
 /**
- * @Created：2018-08-27
+ * @Created：2018-08-28
  * @Author Albert
  * @Version: 1.0
  * @Description: FormExtGroup视图类
@@ -19,11 +19,11 @@ public class FormExtGroupVO implements Serializable{
 
     private Date gmtModified;
 
-    private Integer formTypeId;
+    private Long formTypeId;
 
-    private Integer tabId;
+    private Long tabId;
 
-    private Integer nodeId;
+    private Long nodeId;
 
     private String name;
 
@@ -55,27 +55,27 @@ public class FormExtGroupVO implements Serializable{
         this.gmtModified = gmtModified;
     }
 
-    public Integer getFormTypeId() {
+    public Long getFormTypeId() {
         return formTypeId;
     }
 
-    public void setFormTypeId(Integer formTypeId) {
+    public void setFormTypeId(Long formTypeId) {
         this.formTypeId = formTypeId;
     }
 
-    public Integer getTabId() {
+    public Long getTabId() {
         return tabId;
     }
 
-    public void setTabId(Integer tabId) {
+    public void setTabId(Long tabId) {
         this.tabId = tabId;
     }
 
-    public Integer getNodeId() {
+    public Long getNodeId() {
         return nodeId;
     }
 
-    public void setNodeId(Integer nodeId) {
+    public void setNodeId(Long nodeId) {
         this.nodeId = nodeId;
     }
